@@ -4,41 +4,49 @@
     "url": "http://github.com/broofa",
     "email": "robert@broofa.com"
   },
+  "engines": {
+    "node": ">=10.0.0"
+  },
   "bin": {
     "mime": "cli.js"
   },
-  "engines": {
-    "node": ">=4"
-  },
-  "contributors": [
-    {
-      "name": "Benjamin Thomas",
-      "url": "http://github.com/bentomas",
-      "email": "benjamin@benjaminthomas.org"
-    }
-  ],
+  "contributors": [],
   "description": "A comprehensive library for mime-type mapping",
   "license": "MIT",
   "dependencies": {},
   "devDependencies": {
-    "github-release-notes": "0.13.1",
-    "mime-db": "1.31.0",
-    "mime-score": "1.1.0"
+    "benchmark": "*",
+    "chalk": "4.1.2",
+    "eslint": "8.1.0",
+    "mime-db": "1.50.0",
+    "mime-score": "1.2.0",
+    "mime-types": "2.1.33",
+    "mocha": "9.1.3",
+    "runmd": "*",
+    "standard-version": "9.3.2"
   },
+  "files": [
+    "index.js",
+    "lite.js",
+    "Mime.js",
+    "cli.js",
+    "/types"
+  ],
   "scripts": {
-    "prepare": "node src/build.js",
-    "changelog": "gren changelog --tags=all --generate --override",
-    "test": "node src/test.js"
+    "prepare": "node src/build.js && runmd --output README.md src/README_js.md",
+    "release": "standard-version",
+    "benchmark": "node src/benchmark.js",
+    "md": "runmd --watch --output README.md src/README_js.md",
+    "test": "mocha src/test.js"
   },
   "keywords": [
     "util",
     "mime"
   ],
-  "main": "mime.js",
   "name": "mime",
   "repository": {
-    "url": "https://github.com/broofa/node-mime",
+    "url": "https://github.com/broofa/mime",
     "type": "git"
   },
-  "version": "1.6.0"
+  "version": "3.0.0"
 }
