@@ -1,11 +1,10 @@
-const admin = require("firebase-admin");
-const dotenv = require("dotenv");
+import admin from 'firebase-admin';
+import dotenv from 'dotenv';
+import serviceAccount from '../firebase-service-account.json' assert { type: 'json' };
 
 dotenv.config();
 
 // Inisialisasi Firebase Admin SDK menggunakan service account
-const serviceAccount = require("../firebase-service-account.json");
-
 admin.initializeApp({
   credential: admin.credential.cert(serviceAccount),
   databaseURL: process.env.FIREBASE_DATABASE_URL, // Opsional jika menggunakan Realtime Database
@@ -14,4 +13,4 @@ admin.initializeApp({
 // Inisialisasi Firestore
 const db = admin.firestore();
 
-module.exports = db;
+export default db;
