@@ -1,6 +1,12 @@
-const Post = require('../models/postModel');
+import Post from '../models/postModel.js';
 
-exports.getPosts = async (req, res) => {
-    const posts = await Post.find().populate('author');
-    res.json(posts);
+const getPosts = async (req, res) => {
+    try {
+        const posts = await Post.find().populate('author');
+        res.json(posts);
+    } catch (error) {
+        res.status(500).json({ message: 'Error fetching posts', error });
+    }
 };
+
+export default { getPosts };
