@@ -1,6 +1,12 @@
-const User = require('../models/userModel');
+import User from '../models/userModel.js';
 
-exports.getUsers = async (req, res) => {
-    const users = await User.find();
-    res.json(users);
+const getUsers = async (req, res) => {
+    try {
+        const users = await User.find();
+        res.json(users);
+    } catch (error) {
+        res.status(500).json({ message: 'Error fetching users', error });
+    }
 };
+
+export default { getUsers };
