@@ -1,9 +1,10 @@
-const express = require('express');
+import express from 'express';
+import userController from '../controllers/userController.js';
+import postController from '../controllers/postController.js';
+
 const router = express.Router();
-const userController = require('../controllers/userController');
-const postController = require('../controllers/postController');
 
 router.get('/users', userController.getUsers);
 router.get('/posts', postController.getPosts);
 
-module.exports = router;
+export default router;
