@@ -1,7 +1,8 @@
-const express = require("express");
+import express from 'express';
+import db from '../config/firebase.js';
+import articleController from '../controllers/articleController.js';
+
 const router = express.Router();
-const db = require("../config/firebase");
-const articleController = require("../controllers/articleController");
 
 // Rute API
 router.post("/", articleController.addArticle);
@@ -10,7 +11,7 @@ router.get("/:slug", articleController.getArticleBySlug); // Menggunakan route i
 router.put("/:id", articleController.updateArticle);
 router.delete("/:id", articleController.deleteArticle);
 
-// Rute untuk tampilan artikel berdasarkan slug (Dihapus, karena sudah digabung ke route utama)
+// Rute untuk tampilan artikel berdasarkan slug
 router.get("/view/:slug", async (req, res) => {
   try {
     const q = db.collection("articles").where("slug", "==", req.params.slug);
@@ -27,4 +28,4 @@ router.get("/view/:slug", async (req, res) => {
   }
 });
 
-module.exports = router;
+export default router;
