@@ -1,8 +1,29 @@
-const express = require('express');
+import express from 'express';
+
 const router = express.Router();
 
+// Route untuk halaman utama
 router.get('/', (req, res) => {
-  res.render('index', { title: 'Home' });
+  res.render('index', { 
+    title: 'Home', 
+    message: 'Welcome to My Express App!' 
+  });
 });
 
-module.exports = router;
+// Route untuk halaman about (contoh tambahan)
+router.get('/about', (req, res) => {
+  res.render('about', { 
+    title: 'About', 
+    message: 'This is the about page.' 
+  });
+});
+
+// Route untuk halaman kontak (contoh tambahan)
+router.get('/contact', (req, res) => {
+  res.render('contact', { 
+    title: 'Contact', 
+    message: 'You can reach us at contact@example.com.' 
+  });
+});
+
+export default router;
